split out views into separate files
