+++ agileHR/views/__init__.py
@@ -1,0 +1,5 @@
+from .index_views import *
+from .employee_views import *
+from .department_views import *
+from .training_views import *
+from .computer_views import *
