+++ agileHR/views/index_views.py
@@ -1,0 +1,11 @@
+from datetime import datetime
+from django.shortcuts import get_object_or_404, render
+from django.http import HttpResponse, HttpResponseRedirect
+from django.urls import reverse
+from django.contrib import messages
+from agileHR.models import *
+
+
+def index(request):
+    context = {}
+    return render(request, "agileHR/index.html", context)
